feat(quiz): unify entire flow into single /quiz route with seamless step morphing
Collapse /content-selection, /question-count, /questionnaire, and /results
into a single /quiz route whose card surface morphs through all four steps
plus the AI-generation loader without route navigations. Each step swap is
an AnimatePresence cross-fade with the parent's `layout` transition driving
smooth height interpolation between bodies of different sizes.

Side fixes carried in the same refactor:
- Recent picks: add the missing music rail; align movie/book accents to the
  canonical type-accent palette (movies=amber, books=emerald, music=rose).
- Latest pick spotlight + content-selection: per-type eyebrow accent.
- Library + History: grid/list view choice persists across pages via a new
  usePersistedViewMode hook; stat tiles always live above the filter row so
  the 5-tile layout doesn't break the status segmented control.
- Results headline: ships an array of 10 phrases; one is picked per visit.
- Results "Your picks" eyebrow now uses the type-accent color.
