+++ public/sitemap.xml
@@ -11,19 +11,9 @@
     <priority>0.8</priority>
   </url>
   <url>
-    <loc>https://smartadvisor.live/select-content</loc>
+    <loc>https://smartadvisor.live/quiz</loc>
     <changefreq>monthly</changefreq>
     <priority>0.7</priority>
-  </url>
-  <url>
-    <loc>https://smartadvisor.live/question-count</loc>
-    <changefreq>monthly</changefreq>
-    <priority>0.6</priority>
-  </url>
-  <url>
-    <loc>https://smartadvisor.live/questionnaire</loc>
-    <changefreq>monthly</changefreq>
-    <priority>0.6</priority>
   </url>
   <url>
     <loc>https://smartadvisor.live/results</loc>
