+++ src/app/results/page.tsx
@@ -1,1519 +1,35 @@
 "use client";
-import { useState, useEffect, useRef, useCallback, useMemo } from "react";
+
+import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import Image from "next/image";
-import { useTranslations, useMessages } from "next-intl";
-import {
-  RefreshCw,
-  Heart,
-  Sparkles,
-  Star,
-  BookOpen,
-  Film,
-  Music,
-  ArrowRight,
-  Share2,
-  RotateCcw,
-  Copy,
-  Mail,
-  ChevronDown,
-} from "lucide-react";
-import { motion, AnimatePresence } from "motion/react";
-import { useAuth } from "@/features/auth/hooks/use-auth";
-import { useRequireAuth } from "@/features/auth/hooks/use-require-auth";
-import { enhancedRecommendationsService } from "@/features/recommendations/services/enhanced-recommendations-service";
-import { databaseService } from "@/features/recommendations/services/database-service";
-import { Recommendation } from "@/features/recommendations/types/recommendation";
-import {
-  deriveMatchScore,
-  MATCH_TONE_CLASSES,
-} from "@/features/recommendations/utils/match-score";
-import { SafeLocalStorage } from "@/utils/localStorage";
-import {
-  useQuizStore,
-  type ContentType,
-} from "@/features/quiz/store/quiz-store";
-import { getAccentTone } from "@/features/quiz/utils/content-accent";
-import { PillButton } from "@/components/ui/pill-button";
-import { HoverBorderGradient } from "@/components/ui/hover-border-gradient";
-import { SectionHeader } from "@/components/section-header";
-import { LogToLibraryButton } from "@/features/library/components/log-to-library-button";
-import { WishlistButton } from "@/features/library/components/wishlist-button";
-import { libraryService } from "@/features/library/services/library-service";
-import type { LibraryStatus } from "@/features/library/types/library";
+import { useTranslations } from "next-intl";
+
 import { PageLoader } from "@/components/ui/loader";
-import { AppNavbar } from "@/components/app-navbar";
-import { TrailerEmbed } from "@/components/trailer-embed";
-import { MusicPreview } from "@/components/music-preview";
-import { cn } from "@/lib/utils";
-import { toast } from "sonner";
+import { useQuizStore } from "@/features/quiz/store/quiz-store";
 
-const safeStringify = (obj: unknown): string => {
-  const seen = new WeakSet();
-
-  try {
-    return JSON.stringify(obj, (_key, value) => {
-      if (typeof value === "object" && value !== null) {
-        if (seen.has(value)) {
-          return "[Circular Reference]";
-        }
-        seen.add(value);
-      }
-      if (typeof value === "function") return "[Function]";
-      if (typeof value === "symbol") return "[Symbol]";
-      if (value === undefined) return "[Undefined]";
-      return value;
-    });
-  } catch {
-    return `fallback-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-  }
-};
-
-const GENERATED_SESSIONS_KEY = "smart_advisor_generated_sessions";
-
-const getGeneratedSessions = (): Set<string> => {
-  const stored = SafeLocalStorage.getJSON<string[]>(GENERATED_SESSIONS_KEY, []);
-  return new Set(stored);
-};
-
-const saveGeneratedSession = (sessionId: string) => {
-  const sessions = getGeneratedSessions();
-  sessions.add(sessionId);
-
-  const sessionsArray = Array.from(sessions);
-  const sessionsToStore =
-    sessionsArray.length > 10 ? sessionsArray.slice(-10) : sessionsArray;
-
-  SafeLocalStorage.setJSON(GENERATED_SESSIONS_KEY, sessionsToStore);
-};
-
-/* ---------- Loading State ----------
- * One skeleton card per upcoming recommendation — same count and same
- * accent color the user is about to see for real. The mapping mirrors the
- * engine's target split in `enhanced-recommendations-service.ts`:
- *   movie       → 3 amber movie skeletons
- *   book        → 3 emerald book skeletons
- *   music       → 3 rose square album skeletons
- *   mix         → 1 each (movie + book + album)
- *   both (legacy) → 2 movies + 1 book
+/**
+ * Legacy `/results` route. The recommendations UI now lives as the final
+ * step inside `/quiz` so the entire flow (content → count → questions →
+ * loader → results) reads as a single morphing card with no route flashes.
  *
- * Each skeleton mirrors the real RecommendationCard shape (poster column
- * + title + meta chips + body lines) so the swap from loader to result is
- * a content fill rather than a layout shift. The rotating status caption
- * at the bottom uses the chosen contentType's bank. */
-type SkeletonFormat = "movie" | "book" | "music";
-
-const SKELETON_TONES: Record<
-  SkeletonFormat,
-  { border: string; surface: string; dot: string; square: boolean }
-> = {
-  movie: {
-    border: "border-amber-200/60 dark:border-amber-500/30",
-    surface:
-      "from-amber-50/80 via-white to-orange-50/60 dark:from-amber-500/10 dark:via-slate-900/40 dark:to-orange-500/10",
-    dot: "bg-amber-500",
-    square: false,
-  },
-  book: {
-    border: "border-emerald-200/60 dark:border-emerald-500/30",
-    surface:
-      "from-emerald-50/80 via-white to-teal-50/60 dark:from-emerald-500/10 dark:via-slate-900/40 dark:to-teal-500/10",
-    dot: "bg-emerald-500",
-    square: false,
-  },
-  music: {
-    border: "border-rose-200/60 dark:border-rose-500/30",
-    surface:
-      "from-rose-50/80 via-white to-pink-50/60 dark:from-rose-500/10 dark:via-slate-900/40 dark:to-pink-500/10",
-    dot: "bg-rose-500",
-    square: true,
-  },
-};
-
-function skeletonFormatsFor(
-  contentType: ContentType | null,
-): SkeletonFormat[] {
-  switch (contentType) {
-    case "movie":
-      return ["movie", "movie", "movie"];
-    case "book":
-      return ["book", "book", "book"];
-    case "music":
-      return ["music", "music", "music"];
-    case "mix":
-      return ["movie", "book", "music"];
-    case "both":
-      return ["movie", "movie", "book"];
-    default:
-      return ["movie", "book", "music"];
-  }
-}
-
-const SkeletonRecCard = ({ format, index }: { format: SkeletonFormat; index: number }) => {
-  const tone = SKELETON_TONES[format];
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 12 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.32, delay: index * 0.08 }}
-      className={cn(
-        "overflow-hidden rounded-3xl border bg-gradient-to-br shadow-sm backdrop-blur-md",
-        tone.border,
-        tone.surface,
-      )}
-    >
-      <div className="grid sm:grid-cols-[168px_1fr]">
-        {tone.square ? (
-          // Music skeleton mirrors the real music card: centered square
-          // cover-within-cover rather than a stretched 1:1 cell.
-          <div className="relative flex items-center justify-center p-5">
-            <div className="relative aspect-square w-full max-w-[136px] overflow-hidden rounded-2xl bg-slate-200/70 shadow-md dark:bg-slate-800/70">
-              <div
-                className="absolute inset-0 animate-pulse bg-gradient-to-br from-slate-200/40 via-white/30 to-slate-200/40 dark:from-slate-800/40 dark:via-slate-900/30 dark:to-slate-800/40"
-                style={{ animationDelay: `${index * 120}ms` }}
-              />
-            </div>
-          </div>
-        ) : (
-          <div className="relative aspect-[2/3] bg-slate-200/60 dark:bg-slate-800/60">
-            <div
-              className="absolute inset-0 animate-pulse bg-gradient-to-br from-slate-200/40 via-white/30 to-slate-200/40 dark:from-slate-800/40 dark:via-slate-900/30 dark:to-slate-800/40"
-              style={{ animationDelay: `${index * 120}ms` }}
-            />
-          </div>
-        )}
-        <div className="flex flex-col gap-4 p-5 sm:p-6">
-          <div className="space-y-3">
-            <div className="h-7 w-3/4 animate-pulse rounded-md bg-slate-200/80 dark:bg-slate-800/80 sm:h-8" />
-            <div className="h-4 w-1/2 animate-pulse rounded-md bg-slate-200/60 dark:bg-slate-800/60" />
-          </div>
-          <div className="flex flex-wrap gap-2">
-            {[0, 1, 2].map((i) => (
-              <div
-                key={i}
-                className="h-6 w-20 animate-pulse rounded-full bg-slate-200/70 dark:bg-slate-800/70"
-                style={{ animationDelay: `${(index * 3 + i) * 80}ms` }}
-              />
-            ))}
-          </div>
-          <div className="space-y-2">
-            <div className="h-4 w-full animate-pulse rounded-md bg-slate-200/60 dark:bg-slate-800/60" />
-            <div className="h-4 w-[92%] animate-pulse rounded-md bg-slate-200/60 dark:bg-slate-800/60" />
-            <div className="h-4 w-[78%] animate-pulse rounded-md bg-slate-200/60 dark:bg-slate-800/60" />
-          </div>
-        </div>
-      </div>
-    </motion.div>
-  );
-};
-
-const ResultsLoadingState = ({
-  step: _step,
-  contentType,
-}: {
-  step: string;
-  contentType: ContentType | null;
-}) => {
-  const messages = useMessages() as {
-    Results?: { loadingMessages?: Record<string, string[]> };
-  };
-  const loadingMessages = useMemo(() => {
-    const bank = messages.Results?.loadingMessages;
-    if (!bank) return [] as string[];
-    const key = contentType === "both" ? "mix" : (contentType ?? "default");
-    return bank[key] ?? bank.default ?? [];
-  }, [messages, contentType]);
-  const [messageIndex, setMessageIndex] = useState(0);
-  const tone = getAccentTone(contentType);
-  const formats = useMemo(
-    () => skeletonFormatsFor(contentType),
-    [contentType],
-  );
-
-  useEffect(() => {
-    if (loadingMessages.length === 0) return;
-    const id = setInterval(() => {
-      setMessageIndex((i) => (i + 1) % loadingMessages.length);
-    }, 3800);
-    return () => clearInterval(id);
-  }, [loadingMessages.length]);
-
-  return (
-    <div className="mx-auto w-full max-w-4xl space-y-5">
-      <div className="grid gap-5">
-        {formats.map((format, i) => (
-          <SkeletonRecCard
-            key={`${format}-${i}`}
-            format={format}
-            index={i}
-          />
-        ))}
-      </div>
-
-      {/* Live status caption — accent dot + rotating per-type message —
-          mirrors the questionnaire skeleton so the visual language is
-          consistent across the flow. */}
-      <div
-        className="flex items-center justify-center gap-2"
-        role="status"
-        aria-live="polite"
-      >
-        <span
-          aria-hidden
-          className={cn("h-2 w-2 animate-pulse rounded-full", tone.dot)}
-        />
-        <AnimatePresence mode="wait">
-          <motion.p
-            key={messageIndex}
-            initial={{ opacity: 0, y: 4 }}
-            animate={{ opacity: 1, y: 0 }}
-            exit={{ opacity: 0, y: -4 }}
-            transition={{ duration: 0.3, ease: "easeOut" }}
-            className={cn(
-              "text-xs font-bold tracking-tight sm:text-sm",
-              tone.text,
-            )}
-          >
-            {loadingMessages[messageIndex] ?? ""}
-          </motion.p>
-        </AnimatePresence>
-      </div>
-    </div>
-  );
-};
-
-/* ---------- Recommendation Card ---------- */
-type WhyTone = {
-  border: string;
-  surface: string;
-  bar: string;
-  icon: string;
-  eyebrow: string;
-};
-
-const WHY_TONES: Record<"movie" | "book" | "music", WhyTone> = {
-  movie: {
-    border: "border-amber-200/60 dark:border-amber-500/30",
-    surface:
-      "from-amber-50/80 via-white to-orange-50/60 dark:from-amber-500/10 dark:via-slate-900/40 dark:to-orange-500/10",
-    bar: "from-amber-400 to-orange-500",
-    icon: "text-amber-600 dark:text-amber-400",
-    eyebrow: "text-amber-700 dark:text-amber-300",
-  },
-  book: {
-    border: "border-emerald-200/60 dark:border-emerald-500/30",
-    surface:
-      "from-emerald-50/80 via-white to-teal-50/60 dark:from-emerald-500/10 dark:via-slate-900/40 dark:to-teal-500/10",
-    bar: "from-emerald-400 to-teal-500",
-    icon: "text-emerald-600 dark:text-emerald-400",
-    eyebrow: "text-emerald-700 dark:text-emerald-300",
-  },
-  music: {
-    border: "border-rose-200/60 dark:border-rose-500/30",
-    surface:
-      "from-rose-50/80 via-white to-pink-50/60 dark:from-rose-500/10 dark:via-slate-900/40 dark:to-pink-500/10",
-    bar: "from-rose-400 to-pink-500",
-    icon: "text-rose-600 dark:text-rose-400",
-    eyebrow: "text-rose-700 dark:text-rose-300",
-  },
-};
-
-const RecommendationCard = ({
-  rec,
-  index,
-  alreadyLogged,
-  libraryStatus,
-  onToggleFavorite,
-}: {
-  rec: Recommendation;
-  index: number;
-  alreadyLogged: boolean;
-  libraryStatus: LibraryStatus | null;
-  onToggleFavorite: (id: string) => void;
-}) => {
-  const tb = useTranslations("Results.body");
-  const [expanded, setExpanded] = useState(false);
-  const [isClamped, setIsClamped] = useState(false);
-  const descRef = useRef<HTMLParagraphElement>(null);
-  const explanation = rec.explanation ?? "";
-  const description = rec.description ?? "";
-  const showDescription = description.length > 0 && description !== explanation;
-  const { score: matchScore, tone: matchTone } = deriveMatchScore(rec);
-  const whyTone = WHY_TONES[rec.type] ?? WHY_TONES.book;
-
-  useEffect(() => {
-    if (expanded || !showDescription) return;
-    const el = descRef.current;
-    if (!el) return;
-    setIsClamped(el.scrollHeight > el.clientHeight + 1);
-  }, [expanded, showDescription, description]);
-
-  return (
-    <motion.article
-      initial={{ opacity: 0, y: 18 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.32, delay: index * 0.06 }}
-      className={cn(
-        "group overflow-hidden rounded-3xl border bg-gradient-to-br shadow-sm backdrop-blur-md transition-all duration-300 hover:-translate-y-0.5 hover:shadow-lg",
-        rec.type === "movie"
-          ? "border-amber-200/60 from-amber-50/80 to-white dark:border-amber-500/30 dark:from-amber-500/10 dark:to-slate-900/40"
-          : rec.type === "music"
-            ? "border-rose-200/60 from-rose-50/80 to-white dark:border-rose-500/30 dark:from-rose-500/10 dark:to-slate-900/40"
-            : "border-emerald-200/60 from-emerald-50/80 to-white dark:border-emerald-500/30 dark:from-emerald-500/10 dark:to-slate-900/40",
-      )}
-    >
-      {/* Mobile: poster + header on top; rich body below spans full width.
-          Desktop (sm+): grid with full poster column on the left. */}
-      <div className="flex gap-3 p-3 sm:hidden">
-        <div
-          className={cn(
-            "relative w-24 shrink-0 overflow-hidden rounded-xl bg-slate-200 dark:bg-slate-800",
-            rec.type === "music" ? "aspect-square" : "aspect-[2/3]",
-          )}
-        >
-          {rec.poster_url ? (
-            <Image
-              src={rec.poster_url}
-              alt={rec.title}
-              fill
-              sizes="96px"
-              className="object-cover"
-            />
-          ) : (
-            <div className="flex h-full w-full items-center justify-center text-slate-400">
-              {rec.type === "movie" ? (
-                <Film size={20} />
-              ) : rec.type === "music" ? (
-                <Music size={20} />
-              ) : (
-                <BookOpen size={20} />
-              )}
-            </div>
-          )}
-          <div
-            className={cn(
-              "absolute left-1 top-1 rounded-full px-1.5 py-0.5 text-[9px] font-bold shadow-sm backdrop-blur-sm",
-              MATCH_TONE_CLASSES[matchTone],
-            )}
-          >
-            {tb("matchPercent", { score: matchScore })}
-          </div>
-        </div>
-
-        <div className="flex min-w-0 flex-1 flex-col">
-          <div className="flex items-start justify-between gap-2">
-            <div className="min-w-0 flex-1">
-              <span className="inline-flex items-center gap-1 rounded-full bg-slate-100 px-1.5 py-0.5 text-[9px] font-bold uppercase tracking-wider text-slate-600 dark:bg-slate-800 dark:text-slate-300">
-                {rec.type === "movie" ? (
-                  <Film size={9} />
-                ) : rec.type === "music" ? (
-                  <Music size={9} />
-                ) : (
-                  <BookOpen size={9} />
-                )}
-                {rec.type}
-              </span>
-              <h2 className="mt-1 text-base font-black leading-tight tracking-tight">
-                {rec.title}
-              </h2>
-              <p className="mt-0.5 line-clamp-1 text-xs text-slate-500 dark:text-slate-400">
-                {rec.artist
-                  ? tb("byArtist", { artist: rec.artist })
-                  : rec.author
-                    ? tb("byAuthor", { author: rec.author })
-                    : rec.director
-                      ? tb("byDirector", { director: rec.director })
-                      : ""}
-                {rec.year ? ` · ${rec.year}` : ""}
-              </p>
-            </div>
-            <button
-              type="button"
-              onClick={() => onToggleFavorite(rec.id)}
-              aria-label={
-                rec.is_favorited
-                  ? tb("favoriteRemove", { title: rec.title })
-                  : tb("favoriteAdd", { title: rec.title })
-              }
-              className={cn(
-                "shrink-0 rounded-full p-1.5 transition-all active:scale-[0.95]",
-                rec.is_favorited
-                  ? "bg-rose-500 text-white shadow-md"
-                  : "bg-slate-100 text-slate-500 dark:bg-slate-800 dark:text-slate-400",
-              )}
-            >
-              <Heart
-                size={13}
-                fill={rec.is_favorited ? "currentColor" : "none"}
-              />
-            </button>
-          </div>
-          <div className="mt-2 flex flex-wrap items-center gap-1.5">
-            {typeof rec.rating === "number" && (
-              <span className="inline-flex items-center gap-1 rounded-full bg-amber-50 px-1.5 py-0.5 text-[10px] font-bold text-amber-700 dark:bg-amber-900/30 dark:text-amber-400">
-                <Star size={10} className="fill-current" />
-                {rec.rating}
-              </span>
-            )}
-            {rec.genres?.slice(0, 2).map((g) => (
-              <span
-                key={g}
-                className="rounded-full bg-slate-100 px-1.5 py-0.5 text-[10px] font-semibold text-slate-600 dark:bg-slate-800 dark:text-slate-300"
-              >
-                {g}
-              </span>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      {/* Mobile-only body (below the header strip) */}
-      <div className="space-y-3 px-3 pb-3 sm:hidden">
-        {explanation && (
-          <div
-            className={cn(
-              "relative overflow-hidden rounded-xl border bg-gradient-to-br p-3",
-              whyTone.border,
-              whyTone.surface,
-            )}
-          >
-            <span
-              aria-hidden="true"
-              className={cn(
-                "absolute inset-y-0 left-0 w-1 bg-gradient-to-b",
-                whyTone.bar,
-              )}
-            />
-            <div className="flex items-center gap-1.5 pl-2">
-              <Sparkles size={12} className={whyTone.icon} />
-              <p
-                className={cn(
-                  "text-[10px] font-black uppercase tracking-[0.16em]",
-                  whyTone.eyebrow,
-                )}
-              >
-                {tb("whyThisPick")}
-              </p>
-            </div>
-            <p className="mt-1.5 pl-2 text-sm leading-relaxed text-slate-700 dark:text-slate-200">
-              {explanation}
-            </p>
-          </div>
-        )}
-
-        <div className="flex items-center justify-between gap-2">
-          <div className="flex items-center gap-2">
-            <LogToLibraryButton
-              medium={rec.type}
-              title={rec.title}
-              creator={rec.artist ?? rec.author ?? rec.director ?? null}
-              year={rec.year ?? null}
-              poster_url={rec.poster_url ?? null}
-              source_recommendation_id={rec.id}
-              initialLogged={alreadyLogged}
-              variant="compact"
-            />
-            <WishlistButton
-              medium={rec.type}
-              title={rec.title}
-              creator={rec.artist ?? rec.author ?? rec.director ?? null}
-              year={rec.year ?? null}
-              poster_url={rec.poster_url ?? null}
-              source_recommendation_id={rec.id}
-              initialStatus={libraryStatus}
-              variant="compact"
-            />
-          </div>
-        </div>
-        {rec.type === "music" && rec.preview_url && (
-          <MusicPreview previewUrl={rec.preview_url} />
-        )}
-
-        {showDescription && (
-          <div className="border-t border-slate-100 pt-3 dark:border-slate-800">
-            <p className="mb-1 text-[10px] font-black uppercase tracking-[0.16em] text-slate-400 dark:text-slate-500">
-              {tb("about")}
-            </p>
-            <p
-              className="overflow-hidden text-sm leading-relaxed text-slate-600 dark:text-slate-400"
-              style={
-                expanded
-                  ? undefined
-                  : {
-                      display: "-webkit-box",
-                      WebkitLineClamp: 3,
-                      WebkitBoxOrient: "vertical",
-                    }
-              }
-            >
-              {description}
-            </p>
-            {description.length > 180 && (
-              <button
-                type="button"
-                onClick={() => setExpanded((prev) => !prev)}
-                className="mt-1.5 inline-flex items-center gap-1 rounded-full px-2 py-0.5 text-[10px] font-bold uppercase tracking-wider text-indigo-600 dark:text-indigo-400"
-              >
-                {expanded ? tb("showLess") : tb("showMore")}
-                <ChevronDown
-                  size={10}
-                  className={cn(
-                    "transition-transform duration-200",
-                    expanded && "rotate-180",
-                  )}
-                />
-              </button>
-            )}
-          </div>
-        )}
-
-        {rec.type !== "music" && (
-          <TrailerEmbed
-            type={rec.type}
-            title={rec.title}
-            year={rec.year ?? null}
-            author={rec.author ?? null}
-          />
-        )}
-      </div>
-
-      <div className="hidden sm:grid sm:grid-cols-[168px_1fr]">
-        {/* Poster — full-bleed for tall movie/book posters; for music
-            albums we wrap the square cover in a tinted column so it sits
-            as a centered, properly-aspected card-within-card rather than
-            getting vertically cropped by the grid row stretching to match
-            the content side. */}
-        {rec.type === "music" ? (
-          <div
-            className={cn(
-              "relative flex items-center justify-center p-5 bg-gradient-to-br",
-              whyTone.surface,
-            )}
-          >
-            <div className="relative aspect-square w-full max-w-[136px] overflow-hidden rounded-2xl bg-slate-200 shadow-lg shadow-rose-500/10 dark:bg-slate-800">
-              {rec.poster_url ? (
-                <Image
-                  src={rec.poster_url}
-                  alt={rec.title}
-                  fill
-                  sizes="136px"
-                  className="object-cover transition-transform duration-500 group-hover:scale-105"
-                />
-              ) : (
-                <div className="flex h-full w-full items-center justify-center text-slate-400">
-                  <Music size={32} />
-                </div>
-              )}
-            </div>
-            <div className="absolute left-3 top-3 inline-flex items-center gap-1 rounded-full bg-rose-500 px-2.5 py-1 text-[10px] font-bold uppercase tracking-wider text-white shadow-sm">
-              <Music size={10} />
-              {rec.type}
-            </div>
-            <div
-              className={cn(
-                "absolute right-3 top-3 rounded-full px-2.5 py-1 text-[10px] font-bold shadow-sm backdrop-blur-sm",
-                MATCH_TONE_CLASSES[matchTone],
-              )}
-            >
-              {tb("matchSuffix", { score: matchScore })}
-            </div>
-          </div>
-        ) : (
-          <div className="relative aspect-[2/3] overflow-hidden bg-slate-200 dark:bg-slate-800">
-            {rec.poster_url ? (
-              <Image
-                src={rec.poster_url}
-                alt={rec.title}
-                fill
-                sizes="168px"
-                className="object-cover transition-transform duration-500 group-hover:scale-105"
-              />
-            ) : (
-              <div className="flex h-full w-full items-center justify-center text-slate-400">
-                {rec.type === "movie" ? (
-                  <Film size={32} />
-                ) : (
-                  <BookOpen size={32} />
-                )}
-              </div>
-            )}
-            <div className="absolute left-3 top-3 inline-flex items-center gap-1 rounded-full bg-black/60 px-2.5 py-1 text-[10px] font-bold uppercase tracking-wider text-white backdrop-blur-sm">
-              {rec.type === "movie" ? (
-                <Film size={10} />
-              ) : (
-                <BookOpen size={10} />
-              )}
-              {rec.type}
-            </div>
-            <div
-              className={cn(
-                "absolute right-3 top-3 rounded-full px-2.5 py-1 text-[10px] font-bold shadow-sm backdrop-blur-sm",
-                MATCH_TONE_CLASSES[matchTone],
-              )}
-            >
-              {tb("matchSuffix", { score: matchScore })}
-            </div>
-          </div>
-        )}
-
-        {/* Content */}
-        <div className="flex flex-col p-5 sm:p-6">
-          <div className="flex items-start justify-between gap-4">
-            <div className="min-w-0 flex-1">
-              <h2 className="text-xl font-black tracking-tight sm:text-2xl">
-                {rec.title}
-              </h2>
-              <p className="mt-1 text-sm text-slate-500 dark:text-slate-400">
-                {rec.artist
-                  ? tb("byArtist", { artist: rec.artist })
-                  : rec.author
-                    ? tb("byAuthor", { author: rec.author })
-                    : rec.director
-                      ? tb("byDirector", { director: rec.director })
-                      : ""}
-                {rec.year ? ` · ${rec.year}` : ""}
-              </p>
-            </div>
-            <div className="flex shrink-0 items-center gap-2">
-              <LogToLibraryButton
-                medium={rec.type}
-                title={rec.title}
-                creator={rec.artist ?? rec.author ?? rec.director ?? null}
-                year={rec.year ?? null}
-                poster_url={rec.poster_url ?? null}
-                source_recommendation_id={rec.id}
-                initialLogged={alreadyLogged}
-              />
-              <WishlistButton
-                medium={rec.type}
-                title={rec.title}
-                creator={rec.artist ?? rec.author ?? rec.director ?? null}
-                year={rec.year ?? null}
-                poster_url={rec.poster_url ?? null}
-                source_recommendation_id={rec.id}
-                initialStatus={libraryStatus}
-              />
-              <button
-                type="button"
-                onClick={() => onToggleFavorite(rec.id)}
-                aria-label={
-                  rec.is_favorited
-                    ? tb("favoriteRemove", { title: rec.title })
-                    : tb("favoriteAdd", { title: rec.title })
-                }
-                className={cn(
-                  "rounded-full p-2 transition-all active:scale-[0.95]",
-                  rec.is_favorited
-                    ? "bg-rose-500 text-white shadow-md"
-                    : "bg-slate-100 text-slate-500 hover:bg-slate-200 dark:bg-slate-800 dark:text-slate-400 dark:hover:bg-slate-700",
-                )}
-              >
-                <Heart
-                  size={16}
-                  fill={rec.is_favorited ? "currentColor" : "none"}
-                />
-              </button>
-            </div>
-          </div>
-
-          {(typeof rec.rating === "number" || rec.genres?.length) && (
-            <div className="mt-3 flex flex-wrap items-center gap-2">
-              {typeof rec.rating === "number" && (
-                <span className="inline-flex items-center gap-1 rounded-full bg-amber-50 px-2.5 py-1 text-xs font-bold text-amber-700 dark:bg-amber-900/30 dark:text-amber-400">
-                  <Star size={12} className="fill-current" />
-                  {rec.rating}
-                </span>
-              )}
-              {rec.genres?.slice(0, 4).map((g) => (
-                <span
-                  key={g}
-                  className="rounded-full bg-slate-100 px-2.5 py-1 text-xs font-semibold text-slate-600 dark:bg-slate-800 dark:text-slate-300"
-                >
-                  {g}
-                </span>
-              ))}
-            </div>
-          )}
-
-          {explanation && (
-            <div
-              className={cn(
-                "relative mt-4 overflow-hidden rounded-2xl border bg-gradient-to-br p-4",
-                whyTone.border,
-                whyTone.surface,
-              )}
-            >
-              <span
-                aria-hidden="true"
-                className={cn(
-                  "absolute inset-y-0 left-0 w-1 bg-gradient-to-b",
-                  whyTone.bar,
-                )}
-              />
-              <div className="flex items-center gap-2 pl-2">
-                <Sparkles size={14} className={whyTone.icon} />
-                <p
-                  className={cn(
-                    "text-[11px] font-black uppercase tracking-[0.16em]",
-                    whyTone.eyebrow,
-                  )}
-                >
-                  {tb("whyThisPick")}
-                </p>
-              </div>
-              <p className="mt-2 pl-2 text-[15px] leading-relaxed text-slate-700 dark:text-slate-200">
-                {explanation}
-              </p>
-            </div>
-          )}
-
-          {showDescription && (
-            <div className="mt-4 border-t border-slate-100 pt-4 dark:border-slate-800">
-              <p className="mb-1.5 text-[11px] font-black uppercase tracking-[0.16em] text-slate-400 dark:text-slate-500">
-                {tb("about")}
-              </p>
-              <p
-                ref={descRef}
-                className="overflow-hidden text-sm leading-relaxed text-slate-600 dark:text-slate-400"
-                style={
-                  expanded
-                    ? undefined
-                    : {
-                        display: "-webkit-box",
-                        WebkitLineClamp: 3,
-                        WebkitBoxOrient: "vertical",
-                      }
-                }
-              >
-                {description}
-              </p>
-              {isClamped && (
-                <button
-                  type="button"
-                  onClick={() => setExpanded((prev) => !prev)}
-                  aria-label={expanded ? tb("showLess") : tb("showMore")}
-                  aria-expanded={expanded}
-                  className="mt-2 inline-flex items-center gap-1 rounded-full px-2.5 py-1 text-[11px] font-bold uppercase tracking-wider text-indigo-600 transition-colors hover:bg-indigo-50 dark:text-indigo-400 dark:hover:bg-indigo-500/10"
-                >
-                  {expanded ? tb("showLess") : tb("showMore")}
-                  <ChevronDown
-                    size={12}
-                    className={cn(
-                      "transition-transform duration-200",
-                      expanded && "rotate-180",
-                    )}
-                  />
-                </button>
-              )}
-            </div>
-          )}
-
-          {rec.type === "music" && rec.preview_url ? (
-            <div className="mt-4">
-              <MusicPreview previewUrl={rec.preview_url} />
-            </div>
-          ) : rec.type !== "music" ? (
-            <div className="mt-4">
-              <TrailerEmbed
-                type={rec.type}
-                title={rec.title}
-                year={rec.year ?? null}
-                author={rec.author ?? null}
-              />
-            </div>
-          ) : null}
-        </div>
-      </div>
-    </motion.article>
-  );
-};
-
-/* ---------- Main Results Page ---------- */
+ * This page exists only to redirect existing bookmarks / external links —
+ * we forward into `/quiz?step=results` when the in-memory store still has
+ * recs from a recent quiz, and otherwise drop the user at the start.
+ */
 const ResultsPage = () => {
   const router = useRouter();
-  const { user } = useAuth();
-  const { ready } = useRequireAuth();
-  const { contentType, answers, reset } = useQuizStore();
   const tc = useTranslations("Common");
-  const tb = useTranslations("Results.body");
-  const [recommendations, setRecommendations] = useState<Recommendation[]>([]);
-  const [loggedTitleMap, setLoggedTitleMap] = useState<
-    Map<string, LibraryStatus>
-  >(() => new Map());
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [generationStep, setGenerationStep] = useState<string>("");
-  const [showConfirmDialog, setShowConfirmDialog] = useState(false);
-
-  const currentSessionRef = useRef<string | null>(null);
-  const generatedRecommendationsRef = useRef<Recommendation[]>([]);
-  const isInitialLoadRef = useRef(true);
-  const abortControllerRef = useRef<AbortController | null>(null);
-
-  // Generate a unique session ID per quiz run using a stable ref
-  const sessionTimestampRef = useRef<number>(Date.now());
-  const sessionId = (() => {
-    if (!answers || !contentType) return null;
-    try {
-      const serializedAnswers = safeStringify(answers);
-      const userId = user?.id || "anonymous";
-      return `${serializedAnswers}-${contentType}-${userId}-${sessionTimestampRef.current}`;
-    } catch {
-      return `fallback-session-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-    }
-  })();
-
-  const handleGenerateRecommendations = useCallback(async () => {
-    if (!sessionId || !user) return;
-
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-    }
-
-    const abortController = new AbortController();
-    abortControllerRef.current = abortController;
-
-    try {
-      setLoading(true);
-      setError(null);
-
-      if (abortController.signal.aborted) return;
-
-      setGenerationStep(tb("generationStep.analyzing"));
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-      if (abortController.signal.aborted) return;
-
-      setGenerationStep(tb("generationStep.generating"));
-      await new Promise((resolve) => setTimeout(resolve, 500));
-      if (abortController.signal.aborted) return;
-
-      setGenerationStep(tb("generationStep.enhancing"));
-
-      const questionnaireData = {
-        answers,
-        contentType: contentType ?? ("mix" as const),
-        userAge: user.age,
-        userName: user.name,
-      };
-
-      const recs = await enhancedRecommendationsService.retryRecommendation(
-        questionnaireData,
-        user.id,
-      );
-
-      if (abortController.signal.aborted) return;
-
-      setGenerationStep(tb("generationStep.finalizing"));
-      await new Promise((resolve) => setTimeout(resolve, 500));
-      if (abortController.signal.aborted) return;
-
-      generatedRecommendationsRef.current = recs;
-      currentSessionRef.current = sessionId;
-      saveGeneratedSession(sessionId);
-      setRecommendations(recs);
-    } catch (err) {
-      if (abortController.signal.aborted) return;
-      console.error("Error generating recommendations:", err);
-      setError(tb("generationError"));
-    } finally {
-      if (!abortController.signal.aborted) {
-        setLoading(false);
-      }
-      if (abortControllerRef.current === abortController) {
-        abortControllerRef.current = null;
-      }
-    }
-  }, [sessionId, user, answers, contentType, tb]);
-
-  const wasSessionGenerated = useCallback((id: string): boolean => {
-    const generatedSessions = getGeneratedSessions();
-    return generatedSessions.has(id);
-  }, []);
-
-  const handleConfirmGeneration = () => {
-    setShowConfirmDialog(false);
-    handleGenerateRecommendations();
-  };
-
-  const handleCancelGeneration = () => {
-    setShowConfirmDialog(false);
-    router.push("/dashboard");
-  };
+  const { recommendations } = useQuizStore();
 
   useEffect(() => {
-    if (!contentType || !answers?.length || !user || !sessionId) {
-      router.push("/content-selection");
-      return;
-    }
-
-    const isSameSession = currentSessionRef.current === sessionId;
-    const hasExistingRecommendations =
-      generatedRecommendationsRef.current.length > 0;
-
-    if (isSameSession && hasExistingRecommendations) {
-      setRecommendations(generatedRecommendationsRef.current);
-      setLoading(false);
-      return;
-    }
-
-    const sessionWasGenerated = wasSessionGenerated(sessionId);
-
-    if (sessionWasGenerated && isInitialLoadRef.current) {
-      setShowConfirmDialog(true);
-      setLoading(false);
-      isInitialLoadRef.current = false;
-      return;
-    }
-
-    if (!sessionWasGenerated) {
-      handleGenerateRecommendations();
+    if (recommendations.length > 0) {
+      router.replace("/quiz?step=results");
     } else {
-      setLoading(false);
+      router.replace("/quiz");
     }
+  }, [recommendations.length, router]);
 
-    isInitialLoadRef.current = false;
-  }, [
-    contentType,
-    answers,
-    user,
-    sessionId,
-    router,
-    wasSessionGenerated,
-    handleGenerateRecommendations,
-  ]);
-
-  useEffect(() => {
-    if (!ready) return;
-    void libraryService.list().then(({ data }) => {
-      setLoggedTitleMap(
-        new Map(
-          data.map((i) => [`${i.medium}::${i.title.toLowerCase()}`, i.status]),
-        ),
-      );
-    });
-  }, [ready]);
-
-  const libraryStatusOf = useCallback(
-    (rec: Recommendation): LibraryStatus | null =>
-      loggedTitleMap.get(`${rec.type}::${rec.title.toLowerCase()}`) ?? null,
-    [loggedTitleMap],
-  );
-
-  const isAlreadyLogged = useCallback(
-    (rec: Recommendation): boolean => libraryStatusOf(rec) !== null,
-    [libraryStatusOf],
-  );
-
-  useEffect(() => {
-    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-      if (recommendations.length > 0 && !error) {
-        e.preventDefault();
-        e.returnValue = tb("beforeUnload");
-        return tb("beforeUnload");
-      }
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-    return () => window.removeEventListener("beforeunload", handleBeforeUnload);
-  }, [recommendations.length, error, tb]);
-
-  useEffect(() => {
-    return () => {
-      if (abortControllerRef.current) {
-        abortControllerRef.current.abort();
-        abortControllerRef.current = null;
-      }
-
-      currentSessionRef.current = null;
-      generatedRecommendationsRef.current = [];
-      isInitialLoadRef.current = true;
-    };
-  }, []);
-
-  const handleGetAnother = () => {
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-      abortControllerRef.current = null;
-    }
-
-    currentSessionRef.current = null;
-    generatedRecommendationsRef.current = [];
-    isInitialLoadRef.current = true;
-    reset();
-    router.push("/content-selection");
-  };
-
-  const handleToggleFavorite = async (recommendationId: string) => {
-    try {
-      if (!recommendationId) return;
-
-      const { error: toggleError } =
-        await databaseService.toggleFavorite(recommendationId);
-      if (toggleError) {
-        toast.error(tb("shareToasts.favoriteFailed"));
-      } else {
-        const rec = recommendations.find((r) => r.id === recommendationId);
-        toast.success(
-          rec?.is_favorited
-            ? tb("shareToasts.favoriteRemoved")
-            : tb("shareToasts.favoriteAdded"),
-        );
-        const updatedRecs = recommendations.map((r) =>
-          r.id === recommendationId
-            ? { ...r, is_favorited: !r.is_favorited }
-            : r,
-        );
-        setRecommendations(updatedRecs);
-        generatedRecommendationsRef.current = updatedRecs;
-      }
-    } catch (err) {
-      console.error("Error toggling favorite:", err);
-    }
-  };
-
-  const handleRetry = () => {
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-      abortControllerRef.current = null;
-    }
-
-    currentSessionRef.current = null;
-    generatedRecommendationsRef.current = [];
-    isInitialLoadRef.current = true;
-    handleGenerateRecommendations();
-  };
-
-  const [showShareMenu, setShowShareMenu] = useState(false);
-
-  const buildShareText = () => {
-    const mRecs = recommendations.filter((r) => r.type === "movie");
-    const bRecs = recommendations.filter((r) => r.type === "book");
-    const musicRecsLocal = recommendations.filter((r) => r.type === "music");
-    const lines: string[] = [tb("shareText.title"), ""];
-
-    if (mRecs.length > 0) {
-      lines.push(tb("shareText.movies"));
-      mRecs.forEach((r) => {
-        const detail = [
-          r.director &&
-            tb("shareText.directorPrefix", { director: r.director }),
-          r.year,
-        ]
-          .filter(Boolean)
-          .join(", ");
-        lines.push(`  ${r.title}${detail ? ` (${detail})` : ""}`);
-        if (r.genres?.length)
-          lines.push(
-            `  ${tb("shareText.genres", { list: r.genres.join(", ") })}`,
-          );
-      });
-      lines.push("");
-    }
-
-    if (bRecs.length > 0) {
-      lines.push(tb("shareText.books"));
-      bRecs.forEach((r) => {
-        const detail = [
-          r.author && tb("shareText.byAuthorShort", { author: r.author }),
-          r.year,
-        ]
-          .filter(Boolean)
-          .join(", ");
-        lines.push(`  ${r.title}${detail ? ` (${detail})` : ""}`);
-        if (r.genres?.length)
-          lines.push(
-            `  ${tb("shareText.genres", { list: r.genres.join(", ") })}`,
-          );
-      });
-      lines.push("");
-    }
-
-    if (musicRecsLocal.length > 0) {
-      lines.push(tb("shareText.music"));
-      musicRecsLocal.forEach((r) => {
-        const detail = [
-          r.artist && tb("shareText.byArtistShort", { artist: r.artist }),
-          r.year,
-        ]
-          .filter(Boolean)
-          .join(", ");
-        lines.push(`  ${r.title}${detail ? ` (${detail})` : ""}`);
-        if (r.genres?.length)
-          lines.push(
-            `  ${tb("shareText.genres", { list: r.genres.join(", ") })}`,
-          );
-      });
-      lines.push("");
-    }
-
-    lines.push(tb("shareText.footer"));
-    return lines.join("\n");
-  };
-
-  const handleShare = async () => {
-    const text = buildShareText();
-
-    if (navigator.share) {
-      try {
-        await navigator.share({ title: tb("shareText.title"), text });
-        return;
-      } catch {
-        // User cancelled — fall through to share menu
-      }
-    }
-
-    setShowShareMenu((prev) => !prev);
-  };
-
-  const handleCopyToClipboard = async () => {
-    try {
-      await navigator.clipboard.writeText(buildShareText());
-      toast.success(tb("shareToasts.copySuccess"));
-    } catch {
-      toast.error(tb("shareToasts.copyFailed"));
-    }
-    setShowShareMenu(false);
-  };
-
-  const handleShareTwitter = () => {
-    const titles = recommendations.map((r) => r.title).join(", ");
-    const text = encodeURIComponent(tb("shareTwitter", { titles }));
-    window.open(`https://x.com/intent/tweet?text=${text}`, "_blank");
-    setShowShareMenu(false);
-  };
-
-  const handleShareEmail = () => {
-    const subject = encodeURIComponent(tb("shareEmailSubject"));
-    const body = encodeURIComponent(buildShareText());
-    window.open(`mailto:?subject=${subject}&body=${body}`, "_blank");
-    setShowShareMenu(false);
-  };
-
-  if (!ready) {
-    return <PageLoader text={tc("loading")} />;
-  }
-
-  if (!contentType || !answers?.length || !user) {
-    return <PageLoader text={tc("loading")} />;
-  }
-
-  const topBar = <AppNavbar />;
-
-  if (showConfirmDialog) {
-    return (
-      <div className="min-h-screen w-full bg-slate-50 text-slate-900 antialiased transition-colors duration-300 dark:bg-slate-950 dark:text-slate-100">
-        {topBar}
-        <main className="px-4 pb-20 pt-32 md:pt-36 sm:px-6">
-          <div className="mx-auto max-w-xl rounded-3xl border border-violet-300/60 bg-gradient-to-br from-violet-50/80 via-fuchsia-50/40 to-rose-50/60 p-8 text-center shadow-sm backdrop-blur-md dark:border-violet-500/40 dark:from-violet-500/15 dark:via-fuchsia-500/10 dark:to-rose-500/15">
-            <div className="mx-auto mb-4 inline-flex h-14 w-14 items-center justify-center rounded-full bg-indigo-100 dark:bg-indigo-500/20">
-              <RefreshCw
-                size={24}
-                className="text-indigo-600 dark:text-indigo-400"
-              />
-            </div>
-            <h2 className="text-2xl font-black tracking-tight">
-              {tb("confirmDialog.title")}
-            </h2>
-            <p className="mt-2 text-sm text-slate-600 dark:text-slate-400">
-              {tb("confirmDialog.body")}
-            </p>
-            <div className="mt-6 flex items-center justify-center gap-3">
-              <PillButton
-                onClick={handleCancelGeneration}
-                className="border-slate-300 bg-white px-5 py-2.5 text-sm font-semibold text-slate-700 dark:border-slate-700 dark:bg-slate-900/70 dark:text-slate-200"
-              >
-                {tb("confirmDialog.cancel")}
-              </PillButton>
-              <PillButton
-                onClick={handleConfirmGeneration}
-                className="bg-white px-5 py-2.5 text-sm font-black text-black dark:bg-slate-900 dark:text-white"
-              >
-                {tb("confirmDialog.confirm")}
-              </PillButton>
-            </div>
-          </div>
-        </main>
-      </div>
-    );
-  }
-
-  if (loading) {
-    return (
-      <div className="min-h-screen w-full bg-slate-50 text-slate-900 antialiased transition-colors duration-300 dark:bg-slate-950 dark:text-slate-100">
-        {topBar}
-        <main className="px-4 pb-20 pt-32 md:pt-36 sm:px-6">
-          <ResultsLoadingState
-            step={generationStep}
-            contentType={contentType}
-          />
-        </main>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="min-h-screen w-full bg-slate-50 text-slate-900 antialiased transition-colors duration-300 dark:bg-slate-950 dark:text-slate-100">
-        {topBar}
-        <main className="px-4 pb-20 pt-32 md:pt-36 sm:px-6">
-          <div className="mx-auto max-w-xl rounded-3xl border border-violet-300/60 bg-gradient-to-br from-violet-50/80 via-fuchsia-50/40 to-rose-50/60 p-8 text-center shadow-sm backdrop-blur-md dark:border-violet-500/40 dark:from-violet-500/15 dark:via-fuchsia-500/10 dark:to-rose-500/15">
-            <div className="mx-auto mb-5 flex h-24 w-24 items-center justify-center overflow-hidden">
-              <video
-                src="/animations/error-animation.webm"
-                autoPlay
-                loop
-                muted
-                playsInline
-                preload="auto"
-                className="h-full w-full object-contain"
-              />
-            </div>
-            <h2 className="text-2xl font-black tracking-tight">
-              {tb("errorState.title")}
-            </h2>
-            <p className="mt-2 text-sm text-slate-600 dark:text-slate-400">
-              {error}
-            </p>
-            <div className="mt-6 flex items-center justify-center gap-3">
-              <PillButton
-                onClick={handleRetry}
-                className="bg-white px-5 py-2.5 text-sm font-black text-black dark:bg-slate-900 dark:text-white"
-              >
-                {tb("errorState.retry")}
-              </PillButton>
-              <PillButton
-                onClick={() => router.push("/content-selection")}
-                className="border-slate-300 bg-white px-5 py-2.5 text-sm font-semibold text-slate-700 dark:border-slate-700 dark:bg-slate-900/70 dark:text-slate-200"
-              >
-                {tb("errorState.back")}
-              </PillButton>
-            </div>
-          </div>
-        </main>
-      </div>
-    );
-  }
-
-  const movieRecs = recommendations.filter((r) => r.type === "movie");
-  const bookRecs = recommendations.filter((r) => r.type === "book");
-  const musicRecs = recommendations.filter((r) => r.type === "music");
-
-  return (
-    <div className="min-h-screen w-full bg-slate-50 text-slate-900 antialiased transition-colors duration-300 dark:bg-slate-950 dark:text-slate-100">
-      {topBar}
-
-      <main className="px-4 pb-20 pt-28 sm:px-6 md:pt-36">
-        <div className="mx-auto max-w-6xl">
-          <motion.div
-            initial={{ opacity: 0, y: 12 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.35 }}
-            className="mb-6 sm:mb-8"
-          >
-            <p className="text-[10px] font-black uppercase tracking-[0.18em] text-indigo-500 sm:text-xs dark:text-indigo-400">
-              {tb("eyebrow")}
-            </p>
-            <h1 className="mt-1.5 text-2xl font-black tracking-tighter sm:mt-2 sm:text-3xl md:text-4xl lg:text-5xl">
-              {tb("headline")}
-            </h1>
-            <p className="mt-1.5 text-sm text-slate-500 sm:mt-2 dark:text-slate-400">
-              {tb("subhead", { count: recommendations.length })}
-            </p>
-          </motion.div>
-
-          {/* Category sections */}
-          {contentType === "both" || contentType === "mix" ? (
-            <>
-              {movieRecs.length > 0 && (
-                <motion.section
-                  initial={{ opacity: 0, y: 12 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.3, delay: 0.05 }}
-                  className="mb-10"
-                >
-                  <SectionHeader
-                    icon={<Film size={14} />}
-                    eyebrow={tb("sections.movie.eyebrow")}
-                    title={tb("sections.movie.title")}
-                    count={movieRecs.length}
-                    accent="violet"
-                  />
-                  <div className="grid gap-5">
-                    {movieRecs.map((rec, i) => (
-                      <RecommendationCard
-                        key={rec.id}
-                        rec={rec}
-                        index={i}
-                        alreadyLogged={isAlreadyLogged(rec)}
-                        libraryStatus={libraryStatusOf(rec)}
-                        onToggleFavorite={handleToggleFavorite}
-                      />
-                    ))}
-                  </div>
-                </motion.section>
-              )}
-
-              {bookRecs.length > 0 && (
-                <motion.section
-                  initial={{ opacity: 0, y: 12 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.3, delay: 0.1 }}
-                  className="mb-10"
-                >
-                  <SectionHeader
-                    icon={<BookOpen size={14} />}
-                    eyebrow={tb("sections.book.eyebrow")}
-                    title={tb("sections.book.title")}
-                    count={bookRecs.length}
-                    accent="emerald"
-                  />
-                  <div className="grid gap-5">
-                    {bookRecs.map((rec, i) => (
-                      <RecommendationCard
-                        key={rec.id}
-                        rec={rec}
-                        index={i}
-                        alreadyLogged={isAlreadyLogged(rec)}
-                        libraryStatus={libraryStatusOf(rec)}
-                        onToggleFavorite={handleToggleFavorite}
-                      />
-                    ))}
-                  </div>
-                </motion.section>
-              )}
-
-              {musicRecs.length > 0 && (
-                <motion.section
-                  initial={{ opacity: 0, y: 12 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{ duration: 0.3, delay: 0.15 }}
-                  className="mb-10"
-                >
-                  <SectionHeader
-                    icon={<Music size={14} />}
-                    eyebrow={tb("sections.music.eyebrow")}
-                    title={tb("sections.music.title")}
-                    count={musicRecs.length}
-                    accent="rose"
-                  />
-                  <div className="grid gap-5">
-                    {musicRecs.map((rec, i) => (
-                      <RecommendationCard
-                        key={rec.id}
-                        rec={rec}
-                        index={i}
-                        alreadyLogged={isAlreadyLogged(rec)}
-                        libraryStatus={libraryStatusOf(rec)}
-                        onToggleFavorite={handleToggleFavorite}
-                      />
-                    ))}
-                  </div>
-                </motion.section>
-              )}
-            </>
-          ) : (
-            <div className="grid gap-5">
-              {recommendations.map((rec, i) => (
-                <RecommendationCard
-                  key={rec.id}
-                  rec={rec}
-                  index={i}
-                  alreadyLogged={isAlreadyLogged(rec)}
-                  libraryStatus={libraryStatusOf(rec)}
-                  onToggleFavorite={handleToggleFavorite}
-                />
-              ))}
-            </div>
-          )}
-
-          {/* Actions */}
-          <motion.div
-            initial={{ opacity: 0, y: 12 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.3, delay: 0.15 }}
-            className="mt-12 flex flex-col items-center justify-center gap-3 sm:flex-row"
-          >
-            <HoverBorderGradient
-              onClick={handleGetAnother}
-              idleColor="17, 24, 39"
-              darkIdleColor="255, 255, 255"
-              highlightColor="99, 102, 241"
-              darkHighlightColor="129, 140, 248"
-              containerClassName="rounded-full w-full sm:w-auto"
-              className="flex w-full items-center justify-center gap-2 whitespace-nowrap bg-white px-6 py-3 text-sm font-black leading-none tracking-tight text-black sm:w-auto dark:bg-black dark:text-white"
-            >
-              <RotateCcw size={16} />
-              {tb("actions.retake")}
-            </HoverBorderGradient>
-            <PillButton
-              onClick={() => router.push("/history")}
-              className="inline-flex w-full items-center justify-center gap-2 border-slate-300 bg-white px-6 py-3 text-sm font-semibold text-slate-700 sm:w-auto dark:border-slate-700 dark:bg-slate-900/70 dark:text-slate-200"
-            >
-              {tb("actions.viewHistory")}
-              <ArrowRight size={16} />
-            </PillButton>
-            <div className="relative w-full sm:w-auto">
-              <PillButton
-                onClick={handleShare}
-                className="inline-flex w-full items-center justify-center gap-2 border-slate-300 bg-white px-6 py-3 text-sm font-semibold text-slate-700 sm:w-auto dark:border-slate-700 dark:bg-slate-900/70 dark:text-slate-200"
-              >
-                <Share2 size={16} />
-                {tb("actions.share")}
-              </PillButton>
-
-              <AnimatePresence>
-                {showShareMenu && (
-                  <>
-                    <motion.div
-                      initial={{ opacity: 0 }}
-                      animate={{ opacity: 1 }}
-                      exit={{ opacity: 0 }}
-                      className="fixed inset-0 z-40"
-                      onClick={() => setShowShareMenu(false)}
-                    />
-                    <motion.div
-                      initial={{ opacity: 0, y: 8, scale: 0.95 }}
-                      animate={{ opacity: 1, y: 0, scale: 1 }}
-                      exit={{ opacity: 0, y: 8, scale: 0.95 }}
-                      transition={{ duration: 0.15 }}
-                      className="absolute bottom-full left-1/2 z-50 mb-2 w-48 -translate-x-1/2 overflow-hidden rounded-xl border border-slate-200/70 bg-white shadow-lg dark:border-slate-700/60 dark:bg-slate-900"
-                    >
-                      <button
-                        onClick={handleCopyToClipboard}
-                        className="flex w-full items-center gap-2.5 px-4 py-2.5 text-sm font-medium text-slate-700 transition-colors hover:bg-slate-50 dark:text-slate-200 dark:hover:bg-slate-800"
-                      >
-                        <Copy size={14} />
-                        {tb("shareMenu.copy")}
-                      </button>
-                      <button
-                        onClick={handleShareTwitter}
-                        className="flex w-full items-center gap-2.5 px-4 py-2.5 text-sm font-medium text-slate-700 transition-colors hover:bg-slate-50 dark:text-slate-200 dark:hover:bg-slate-800"
-                      >
-                        <svg
-                          viewBox="0 0 24 24"
-                          className="h-3.5 w-3.5 fill-current"
-                        >
-                          <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
-                        </svg>
-                        {tb("shareMenu.shareX")}
-                      </button>
-                      <button
-                        onClick={handleShareEmail}
-                        className="flex w-full items-center gap-2.5 px-4 py-2.5 text-sm font-medium text-slate-700 transition-colors hover:bg-slate-50 dark:text-slate-200 dark:hover:bg-slate-800"
-                      >
-                        <Mail size={14} />
-                        {tb("shareMenu.shareEmail")}
-                      </button>
-                    </motion.div>
-                  </>
-                )}
-              </AnimatePresence>
-            </div>
-          </motion.div>
-        </div>
-      </main>
-    </div>
-  );
+  return <PageLoader text={tc("loading")} />;
 };
 
 export default ResultsPage;
