+++ src/features/quiz/store/quiz-store.ts
@@ -1,5 +1,6 @@
 import { create } from "zustand";
 import { Answer } from "@/features/quiz/types/answer";
+import { Recommendation } from "@/features/recommendations/types/recommendation";
 
 export type ContentType = "movie" | "book" | "music" | "both" | "mix";
 
@@ -19,6 +20,13 @@ interface QuizStore {
   filters: { genres: string[]; moods: string[] };
   setFilters: (filters: { genres: string[]; moods: string[] }) => void;
 
+  // Recommendations are pre-generated inside /quiz's generating step and
+  // handed off here so /results can render them immediately on mount instead
+  // of running its own loader. The store is in-memory only, so a hard reload
+  // of /results clears these and the page falls back to its own gen flow.
+  recommendations: Recommendation[];
+  setRecommendations: (recs: Recommendation[]) => void;
+
   reset: () => void;
 }
 
@@ -28,6 +36,7 @@ const defaultState = {
   answers: [] as Answer[],
   userAge: null as number | null,
   filters: { genres: [] as string[], moods: [] as string[] },
+  recommendations: [] as Recommendation[],
 };
 
 export const useQuizStore = create<QuizStore>((set) => ({
@@ -38,5 +47,6 @@ export const useQuizStore = create<QuizStore>((set) => ({
   setAnswers: (answers) => set({ answers }),
   setUserAge: (userAge) => set({ userAge }),
   setFilters: (filters) => set({ filters }),
+  setRecommendations: (recommendations) => set({ recommendations }),
   reset: () => set(defaultState),
 }));
