+++ src/lib/constants.ts
@@ -8,9 +8,7 @@ export const ROUTES = {
   DASHBOARD: "/dashboard",
   DEMO: "/demo",
   DEMO_RESULTS: "/demo/results",
-  CONTENT_SELECTION: "/content-selection",
-  QUESTION_COUNT: "/question-count",
-  QUESTIONNAIRE: "/questionnaire",
+  QUIZ: "/quiz",
   RESULTS: "/results",
   HISTORY: "/history",
   SETTINGS: "/settings",
